rag-service/requirements.txt
langchain-core
langchain
langchain-community
pymupdf
uvicorn
fastapi
gitpython
pydantic